crm/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
defusedxml==0.8.0rc2
Django==3.2
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.0
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
stripe==12.1.0
typing_extensions==4.13.2
urllib3==2.2.3
gunicorn==21.2.0
